fix google chrome install script

+++ install/chrome/install.sh
@@ -43,8 +43,10 @@ _linux_install() {
 
   case "$OS" in
   Ubuntu)
-    sudo bash -c 'echo "deb [arch=$(dpkg --print-architecture)] http://dl.google.com/linux/chrome/deb/ stable main" >>/etc/apt/sources.list.d/google-chrome.list'
-    sudo wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+    if [ ! -f "/etc/apt/sources.list.d/google-chrome.list" ]; then
+      sudo bash -c 'echo "deb [arch=$(dpkg --print-architecture)] https://dl.google.com/linux/chrome/deb/ stable main" >>/etc/apt/sources.list.d/google-chrome.list'
+      sudo wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+    fi
     sudo apt-get update && sudo apt-get install -y google-chrome-stable
     ;;
   *)
